mainfolder/requirements.txt
asn1crypto==1.2.0
azure-common==1.1.23
azure-core==1.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.0.0
azure-storage-common==2.1.0
cchardet==2.1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
Django==2.2.5
Flask==1.1.1
flask-swagger-ui==3.6.0
html5lib==1.0.1
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
langdetect==1.0.7
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
msrest==0.6.10
networkx==2.4
nltk==3.4.5
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.2
pycparser==2.19
pygpu==0.7.6
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
SpeechRecognition==3.6.3
sqlparse==0.3.0
swagger-client==1.0.0
talon==1.4.4
Theano==1.0.4
urllib3==1.25.6
webencodings==0.5.1
Werkzeug==0.16.0
